Choose archive type per-backup target
Previously, we picked a single archive type for all backup targets. This
results in simpler backups, but means if you add a new backup target you
might have to wait 364 days for your first "yearly" backup to be
created. And really, the first backup of any target should be kept
forever.

Fixes #20.