Moved the backup type to within the backup target loop.  This will determine the backup type for each backup target.  Also added the backup target name to the grep so the return list is limited to the specific backup target.
Added a loop around the cleanup of the backups.  I choose to put the existing routine inside a loop instead of adding it to the backup target loop so the call to --list-archives is only run once instead of after each backup target.  Also added the backup target name to the grep so the return list is limited to the specific backup target.