Always create monthly and yearly archives (fixes #29) (#35)
Previously, the first day of the year would create a yearly archive,
the second day would create a monthly, and the third would create a
daily archive. With this change, yearly, monthly, and daily archives
always exist.

Daily archives are created every time acts is run. If monthly and/or
yearly archives don't exist, the new daily archive is copied to
the other(s).

Copies are based on metadata. That metadata must be read and parsed,
which causes some data transfer. However, that data is quite minimal
and should impose very little cost.

Copies do take some time to run. The time scales with the number of
files in the archive, not with the size of each file. And though
time scales, it scales very slowly. In my tests, copies are faster
than archive creation for all but the smallest archives of 1-5 files.
Copies are substantially faster for large archives.