Comment out most settings in acts.conf.sample
In particular, the example $tarsnap setting only works on Linux, and
it's better that acts error out due to empty $backuptargets than end
up backing up gigs of stuff when users aren't ready for it.

I'm ambivalent about the default $tarsnapbackupoptions. They belong
in tarsnap.conf, and while --humanize-numbers is harmless,
--one-file-system could have a meaningful impact on what's backed up.
But, they are pretty sensible places for users to start if they're
new to tarsnap.