Wrap the archive deletion tarsnap call
With it unwrapped, if you call tarsnap with --print-stats, or if that
is in your tarsnap.conf, you'll get the stats printed after every
archive deletion.

While the information is great to have, if you have many targets,
you get 5n lines of output, with nothing indicating which target it's
referring to. It might be worth exploring a better way to show how
much data got added and/or removed for each target, given that tarsnap
can automatically give that info right back to acts.