Add function to handle common string formatting (#36)
We have a couple different ways we need to format strings. Putting
everything into one function makes it easy to format strings in
ways that look right to our users. This also simplifies the printing
of error output, which required shell-check escapes and multiple
levels of escaped quotes.

It's a bit obtuse to dump all this stuff into one function, but it
gives us the most flexibility.